refactor: remove inner code of deprecated classes
